backend/requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
sentinelsat
shapely
python-dotenv
celery[redis]
redis
rasterio
rio-cogeo
rio-tiler
mercantile
numpy
